requirements.txt
albumentations==1.1.0
faiss-cpu==1.7.1
kornia==0.6.4
matplotlib==3.5.1
numpy==1.22.4
onnx==1.13.0
onnxruntime==1.12.1
opencv_python_headless==4.5.5.64
pandas==1.3.4
Pillow==9.4.0
psutil==5.8.0
pytorch_metric_learning==1.4.0
PyYAML==6.0
requests==2.26.0
scikit_image==0.19.1
scipy==1.7.2
seaborn==0.12.2
sentence_transformers==2.2.2
setuptools==58.0.4
symspellpy==6.7.6
timm==0.6.2.dev0
torch==1.10.0
torchvision==0.11.1
tqdm==4.64.1
umap==0.1.1
memory_profiler==0.61.0
rdflib==6.3.2
